\name{ToxLim-package}
\alias{ToxLim-package}
\alias{ToxLim}
\docType{package}
\title{
  Bioaccumulation modelling using LIM
}
\description{
  Incorporating Ecological Data and Associated Uncertainty in
  Bioaccumulation Modeling
}
\details{
  \tabular{ll}{
    Package: \tab ToxLim\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2009-06-09\cr
    License: \tab GNU Public License 2 or above\cr
  }

  ToxLim predicts the internal concentrations of hydrophobic chemicals in
  aquatic organisms.
}
\author{
  Frederik de Laender (maintainer)

  Karline Soetaert
}
%\references{
%% to be filled
%}
\seealso{
 \code{\link{LimOmega}}, the main function

  \code{LIMlake},  \code{LIMlakeFish},  \code{LIMbarents}, the input food webs.
}

\examples{
\dontrun{
## show examples (see respective help pages for details)
example(LimOmega)
example(LIMlake)
example(LIMbarents)
}
}

\keyword{ package }